+++ src/ri/sl.cpp
@@ -693,7 +693,7 @@ static	TSlFunction		functions[]	=	{
 												TSlLabel	*cLabel	=	new TSlLabel;
 
 												strcpy(cLabel->name,name);
-												cLabel->index	=	currentData.currentOpcodePlace - currentData.code;
+												cLabel->index	=	(int) (currentData.currentOpcodePlace - currentData.code);
 
 												if (reference) {
 													cLabel->next					=	currentData.labelReferences;
@@ -3070,7 +3070,7 @@ CShader	*shaderCreate(const char *shaderName) {
 void	processEscapes(char *str) {
 	int		i,n,j;
 
-	n	=	strlen(str);
+	n	=	(int) strlen(str);
 	for (i=0;i<n;i++) {
 		if (str[i] == '\\') {
 			switch(str[i+1]) {
